requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.8
inflection==0.5.1
jwt==1.3.1
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
sqlparse==0.5.1
stripe==10.11.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
